+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v0.7.x
+- add: parameter `pagination_limit` for LD cb client ([#508](https://github.com/RWTH-EBC/FiLiP/issues/508))
 - update: deprecate support for python 3.9 ([#504](https://github.com/RWTH-EBC/FiLiP/pull/504))
 - update: from ``requests~=2.32.0`` to ``requests~=2.33.0`` in dependency ([#504](https://github.com/RWTH-EBC/FiLiP/pull/504))
 - fix: exception handling in checking broker version ([#503](https://github.com/RWTH-EBC/FiLiP/pull/503))