+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ### v0.7.x
 - add: support Kafka notifications, i.e. ``kafka`` and ``kafkaCustom`` in NGSIv2 subscriptions ([#511](https://github.com/RWTH-EBC/FiLiP/issues/511))
+- add: parameter `pagination_limit` for LD cb client ([#519](https://github.com/RWTH-EBC/FiLiP/pull/519))
+- add: dynamic fiware header for authentification ([#507](https://github.com/RWTH-EBC/FiLiP/pull/507))
+- add: tutorials for using N5GEH cluster ([#507](https://github.com/RWTH-EBC/FiLiP/pull/507))
 - update: deprecate support for python 3.9 ([#504](https://github.com/RWTH-EBC/FiLiP/pull/504))
 - update: from ``requests~=2.32.0`` to ``requests~=2.33.0`` in dependency ([#504](https://github.com/RWTH-EBC/FiLiP/pull/504))
 - fix: exception handling in checking broker version ([#503](https://github.com/RWTH-EBC/FiLiP/pull/503))
+- fix: update regex version for py3.13 compatibility ([#512](https://github.com/RWTH-EBC/FiLiP/pull/512))
+- fix: datatime validation in metadata ([#513](https://github.com/RWTH-EBC/FiLiP/pull/513))
 
 ### v0.7.5
 - fix: allow dynamic header composition in v2 http client ([#493](https://github.com/RWTH-EBC/FiLiP/pull/493))